Refactor PfandautomatView for improved structure and performance
- Reorganized component structure by introducing subcomponents: DepositStats, InstructionsCard, DepositReceiptDialog, and DepositListCard for better maintainability.
- Enhanced barcode scanning and product lookup logic with improved error handling and state management.
- Updated UI elements to ensure accessibility and semantic HTML practices.
- Removed unused imports and animation logic to streamline the component.
- Added detailed JSDoc comments for better documentation and understanding of component functionality.
